separate essential functions into a global codefixer. Start implementing Haskell codefixer
