+++ WLC/image_processing/camera.py
@@ -58,6 +58,6 @@ class Camera:
         to_show = Preferences(show_pic, show_line, show_word, show_character, annotate)
 
         if not image_path:
-            image_path = 'assets/examples/images/fib_2.png'
+            image_path = 'assets/examples/images/python3/fib_2.png'
 
         return self.read_file(image_path, to_show)
